+++ apps/backend/src/media/mod.rs
@@ -1,3 +1,4 @@
+use async_graphql::SimpleObject;
 use sea_orm::FromJsonQueryResult;
 use serde::{Deserialize, Serialize};
 
@@ -29,3 +30,15 @@ pub struct MetadataImage {
 
 #[derive(Clone, Debug, PartialEq, FromJsonQueryResult, Eq, Serialize, Deserialize)]
 pub struct MetadataImages(pub Vec<MetadataImage>);
+
+#[derive(
+    Clone, Debug, PartialEq, FromJsonQueryResult, Eq, Serialize, Deserialize, SimpleObject,
+)]
+pub struct MetadataCreator {
+    pub name: String,
+    pub role: String,
+    pub image_urls: Vec<String>,
+}
+
+#[derive(Clone, Debug, PartialEq, FromJsonQueryResult, Eq, Serialize, Deserialize)]
+pub struct MetadataCreators(pub Vec<MetadataCreator>);