+++ apps/backend/src/media/resolver.rs
@@ -13,12 +13,12 @@ use crate::{
     background::{AfterMediaSeenJob, RecalculateUserSummaryJob, UpdateMetadataJob},
     books::BookSpecifics,
     entities::{
-        collection, creator, genre,
+        collection, genre,
         metadata::{self, Model as MetadataModel},
-        metadata_to_collection, metadata_to_creator, metadata_to_genre,
+        metadata_to_collection, metadata_to_genre,
         prelude::{
-            AudioBook, Book, Collection, Creator, Genre, Metadata, MetadataToCollection, Movie,
-            Podcast, Review, Seen, Show, UserToMetadata, VideoGame,
+            AudioBook, Book, Collection, Genre, Metadata, MetadataToCollection, Movie, Podcast,
+            Review, Seen, Show, UserToMetadata, VideoGame,
         },
         review, seen, user_to_metadata,
         utils::{SeenExtraInformation, SeenPodcastExtraInformation, SeenShowExtraInformation},
@@ -32,12 +32,14 @@ use crate::{
     video_games::VideoGameSpecifics,
 };
 
-use super::{MediaSpecifics, MetadataImage, MetadataImages, LIMIT};
+use super::{
+    MediaSpecifics, MetadataCreator, MetadataCreators, MetadataImage, MetadataImages, LIMIT,
+};
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
 pub struct MediaBaseData {
     pub model: MetadataModel,
-    pub creators: Vec<String>,
+    pub creators: Vec<MetadataCreator>,
     pub poster_images: Vec<String>,
     pub backdrop_images: Vec<String>,
     pub genres: Vec<String>,
@@ -85,7 +87,7 @@ pub struct MediaDetails {
     pub title: String,
     pub description: Option<String>,
     pub lot: MetadataLot,
-    pub creators: Vec<String>,
+    pub creators: Vec<MetadataCreator>,
     pub genres: Vec<String>,
     pub poster_images: Vec<String>,
     pub backdrop_images: Vec<String>,
@@ -102,7 +104,7 @@ pub struct DatabaseMediaDetails {
     pub description: Option<String>,
     #[graphql(name = "type")]
     pub lot: MetadataLot,
-    pub creators: Vec<String>,
+    pub creators: Vec<MetadataCreator>,
     pub genres: Vec<String>,
     pub poster_images: Vec<String>,
     pub backdrop_images: Vec<String>,
@@ -307,14 +309,7 @@ impl MediaService {
             .into_iter()
             .map(|g| g.name)
             .collect();
-        let creators = meta
-            .find_related(Creator)
-            .all(&self.db)
-            .await
-            .unwrap()
-            .into_iter()
-            .map(|c| c.name)
-            .collect();
+        let creators = meta.creators.clone().0;
         let (poster_images, backdrop_images) = self.metadata_images(&meta).await.unwrap();
         Ok(MediaBaseData {
             model: meta,
@@ -681,6 +676,7 @@ impl MediaService {
         Ok(())
     }
 
+    #[allow(clippy::too_many_arguments)]
     pub async fn update_media(
         &self,
         metadata_id: i32,
@@ -688,6 +684,7 @@ impl MediaService {
         description: Option<String>,
         poster_images: Vec<String>,
         backdrop_images: Vec<String>,
+        creators: Vec<MetadataCreator>,
     ) -> Result<()> {
         let mut images = vec![];
         for image in poster_images.into_iter() {
@@ -712,6 +709,7 @@ impl MediaService {
         meta.description = ActiveValue::Set(description);
         meta.images = ActiveValue::Set(MetadataImages(images));
         meta.last_updated_on = ActiveValue::Set(Utc::now());
+        meta.creators = ActiveValue::Set(MetadataCreators(creators));
         meta.save(&self.db).await.ok();
         Ok(())
     }
@@ -727,7 +725,7 @@ impl MediaService {
         publish_date: Option<NaiveDate>,
         poster_images: Vec<String>,
         backdrop_images: Vec<String>,
-        creator_names: Vec<String>,
+        creators: Vec<MetadataCreator>,
         genres: Vec<String>,
     ) -> Result<i32> {
         let mut images = vec![];
@@ -751,30 +749,10 @@ impl MediaService {
             publish_date: ActiveValue::Set(publish_date),
             images: ActiveValue::Set(MetadataImages(images)),
             identifier: ActiveValue::Set(identifier),
+            creators: ActiveValue::Set(MetadataCreators(creators)),
             ..Default::default()
         };
         let metadata = metadata.insert(&self.db).await.unwrap();
-        for name in creator_names.iter() {
-            let creator = if let Some(c) = Creator::find()
-                .filter(creator::Column::Name.eq(name))
-                .one(&self.db)
-                .await
-                .unwrap()
-            {
-                c
-            } else {
-                let c = creator::ActiveModel {
-                    name: ActiveValue::Set(name.to_owned()),
-                    ..Default::default()
-                };
-                c.insert(&self.db).await.unwrap()
-            };
-            let metadata_creator = metadata_to_creator::ActiveModel {
-                metadata_id: ActiveValue::Set(metadata.id),
-                creator_id: ActiveValue::Set(creator.id),
-            };
-            metadata_creator.insert(&self.db).await.ok();
-        }
         for genre in genres {
             let db_genre = if let Some(c) = Genre::find()
                 .filter(genre::Column::Name.eq(&genre))