+++ apps/backend/src/utils.rs
@@ -89,7 +89,16 @@ pub async fn create_app_services(
     let video_games_service = VideoGamesService::new(&db, &igdb_service, &media_service);
     let listennotes_service = ListennotesService::new(&config.podcasts).await;
     let podcasts_service = PodcastsService::new(&db, &listennotes_service, &media_service);
-    let misc_service = MiscService::new(&db, &media_service);
+    let misc_service = MiscService::new(
+        &db,
+        &media_service,
+        &audio_books_service,
+        &books_service,
+        &movies_service,
+        &podcasts_service,
+        &shows_service,
+        &video_games_service,
+    );
     let users_service = UsersService::new(&db, &misc_service, user_created_job);
     let importer_service = ImporterService::new(
         &db,
@@ -219,11 +228,21 @@ fn get_now_timestamp() -> u128 {
 }
 
 pub mod tmdb {
+    use std::{env, fs};
+
     use crate::graphql::PROJECT_NAME;
 
     use super::*;
 
+    #[derive(Debug, Serialize, Deserialize, Clone)]
+    pub struct TmdbCredit {
+        pub name: String,
+        pub known_for_department: String,
+        pub profile_path: Option<String>,
+    }
+
     pub async fn get_client_config(url: &str, access_token: &str) -> (Client, String) {
+        let path = env::temp_dir().join("tmdb-config.json");
         let client: Client = Config::new()
             .add_header(USER_AGENT, format!("{}/{}", AUTHOR, PROJECT_NAME))
             .unwrap()
@@ -240,9 +259,15 @@ pub mod tmdb {
         struct TmdbConfiguration {
             images: TmdbImageConfiguration,
         }
-        let mut rsp = client.get("configuration").await.unwrap();
-        let data: TmdbConfiguration = rsp.body_json().await.unwrap();
-        (client, data.images.secure_base_url)
+        let image_url = if let Some(details) = read_file_to_json::<TmdbConfiguration>(&path) {
+            details.images.secure_base_url
+        } else {
+            let mut rsp = client.get("configuration").await.unwrap();
+            let data: TmdbConfiguration = rsp.body_json().await.unwrap();
+            fs::write(path, serde_json::to_string(&data).unwrap()).ok();
+            data.images.secure_base_url
+        };
+        (client, image_url)
     }
 }